+++ packages/create-block/lib/output.js
@@ -7,19 +7,17 @@ const { render } = require( 'mustache' );
 const { writeFile } = require( 'fs' ).promises;
 
 const writeOutputAsset = async ( inputFile, outputFile, view ) => {
-	const outputFilePath = join( view.slug, 'assets', outputFile );
+	const outputFilePath = join( view.rootDirectory, 'assets', outputFile );
 	await makeDir( dirname( outputFilePath ) );
 	writeFile( outputFilePath, inputFile );
 };
 
 const writeOutputTemplate = async ( inputFile, outputFile, view ) => {
-	const outputFilePath = view.plugin
-		? join( view.slug, outputFile.replace( /\$slug/g, view.slug ) )
-		: outputFile;
-	await makeDir( dirname( outputFilePath ) );
 	// If the rendered template is empty, don't write it. This is how we can conditionally add template files.
 	const renderedFile = render( inputFile, view );
 	if ( renderedFile.trim().length ) {
+		const outputFilePath = join( view.rootDirectory, outputFile );
+		await makeDir( dirname( outputFilePath ) );
 		writeFile( outputFilePath, renderedFile );
 	}
 };